fixed compile errors under ms windows; fixed briefing information loading and display

+++ freesynd/trunk/src/ped.cpp
@@ -652,7 +652,8 @@ bool PedInstance::animate(int elapsed, Mission *mission) {
                             aqt.t_pn.convertPosToXY(&pos_x, &pos_y);
                             int diffx = pos_x - tile_x_ * 256 - off_x_;
                             int diffy = pos_y - tile_y_ * 256 - off_y_;
-                            dist_to_pos_ = (int)sqrt(diffx * diffx + diffy * diffy);
+                            dist_to_pos_ = (int)sqrt((double)
+                                (diffx * diffx + diffy * diffy));
                             moveToDir(mission, elapsed,
                                 -1, pos_x, pos_y,
                                 aqt.multi_var.dist_var.dist,
@@ -690,7 +691,8 @@ bool PedInstance::animate(int elapsed, Mission *mission) {
                             aqt.t_pn.convertPosToXY(&pos_x, &pos_y);
                             int diffx = pos_x - tile_x_ * 256 - off_x_;
                             int diffy = pos_y - tile_y_ * 256 - off_y_;
-                            dist_to_pos_ = (int)sqrt(diffx * diffx + diffy * diffy);
+                            dist_to_pos_ = (int)sqrt((double)
+                                (diffx * diffx + diffy * diffy));
                             if (dist_to_pos_ > 0) {
                                 moveToDir(mission, elapsed,
                                     -1, pos_x, pos_y,
@@ -713,7 +715,8 @@ bool PedInstance::animate(int elapsed, Mission *mission) {
                             aqt.t_smo->convertPosToXY(&pos_x, &pos_y);
                             int diffx = pos_x - tile_x_ * 256 - off_x_;
                             int diffy = pos_y - tile_y_ * 256 - off_y_;
-                            dist_to_pos_ = (int)sqrt(diffx * diffx + diffy * diffy);
+                            dist_to_pos_ = (int)sqrt((double)
+                                (diffx * diffx + diffy * diffy));
                             if (dist_to_pos_ > 0) {
                                 moveToDir(mission, elapsed,
                                     -1, pos_x, pos_y,
